Add initial exploration code for meltdown (sortof)
This commit only adds a signal handler for SIGSEGV and allow code
execution to continue after that the handler was called. It was not
clear to me how acheive this and my first thought was that it would just
work by starting execution after the line that cause the segment fault
but that is not the case.